Remove unnecessary workflow and use step instead
